+++ src/contracts/interface-module.php
@@ -1,0 +1,9 @@
+<?php
+namespace OpenLingua\Contracts;
+
+defined( 'ABSPATH' ) || exit;
+
+interface Module {
+	public static function hooks();
+}
+